feat: added the mock api data and refactor code
