+++ src/app/app.ts
@@ -1,9 +1,11 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ProductGridComponent } from './components/product-grid/product-grid.component';
-import { FooterBarComponent } from './components/footer-bar/footer-bar.component';
-import { CartComponent } from './components/cart/cart.component';
-import { NavBarComponent } from './components/nav-bar/nav-bar.component';
+import {
+  NavBarComponent,
+  FooterBarComponent,
+  CartComponent,
+  ProductGridComponent,
+} from './components';
 
 @Component({
   selector: 'app-root',
