+++ src/app/components/filter-bar/filter-bar.component.html
@@ -15,5 +15,5 @@
       <option value="category2">Category 2</option>
     </select>
   </div>
-  <button class="bg-blue-500 text-white p-2 rounded">Apply Filters</button>
+  <button class="bg-slate-500 text-white p-2 rounded">Apply Filters</button>
 </div>
