+++ src/app/components/nav-bar/nav-bar.component.html
@@ -3,7 +3,7 @@
     <a class="text-white text-2xl font-bold">eLearnio</a>
   </h1>
   <button
-    class="bg-blue-500 text-white px-2 py-1 rounded cursor-pointer"
+    class="bg-slate-500 text-white px-2 py-1 rounded cursor-pointer"
     (click)="toggleCart()"
   >
     Cart
