+++ src/app/components/product-grid/product-grid.component.html
@@ -1,7 +1,7 @@
-<div class="p-4 flex gap-4 flex-col bg-gray-100">
-  <filter-bar />
+<div class="p-4 flex flex-col bg-gray-100">
+  <filter-bar class="my-4" />
   <div
-    class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6"
+    class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-8 my-4"
   >
     @for (product of productStore.products(); track product.id) {
       <product-item
