+++ src/app/components/product-grid/product-item/product-item.component.html
@@ -1,10 +1,10 @@
 <div
-  class="relative p-8 flex flex-col items-center bg-white shadow-md rounded-xl border border-gray-200 overflow-hidden hover:shadow-xl transition-shadow duration-300"
+  class="relative flex flex-col items-center bg-white shadow-md rounded-xl border border-gray-200 overflow-hidden hover:shadow-xl transition-shadow duration-300"
 >
   <img
     [src]="product().image"
     alt="{{ product().title }}"
-    class="w-full h-64 object-fit"
+    class="w-full h-48 object-cover"
   />
   <div class="p-4 flex flex-col justify-start w-full">
     <h2 class="text-xl font-bold truncate">{{ product().title }}</h2>
@@ -12,7 +12,7 @@
   </div>
   <div class="w-full">
     <button
-      class="bg-blue-500 text-white px-4 py-2 rounded w-full"
+      class="bg-slate-500 text-white px-4 py-2 w-full"
       (click)="onAddToCart(product())"
     >
       Add to Cart
