+++ src/app/store/index.ts
@@ -1,1 +1,2 @@
 export * from './product.store';
+export * from './cart.store';
